% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Ricetl.R
\name{id2address}
\alias{id2address}
\title{Extraction Address}
\usage{
id2address(x,fillna=NA)
}
\arguments{
\item{x}{A vector containing 18 Bit ID numbers,\link{id2age}.}

\item{fillna}{The filling value of the missing value,the default value is \emph{NA},\link{id2age}.}
}
\description{
Extraction of age in the resident identity cards of People's Republic of China.
}
\examples{

library(Ricetl)
x <- c('654003198111200241','341881197709275718',
'340503199401248097','469030199406204103','51110019')
id2address(x)
id2address(x,fillna='Error')

}
\author{
Xu Jing
}
