% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Ricetl.R
\name{id2gender}
\alias{id2gender}
\title{Extraction Gender}
\usage{
id2gender(x,fillna=NA)
}
\arguments{
\item{x}{A vector containing 18 Bit ID numbers,\link{id2age}.}

\item{fillna}{The filling value of the missing value,the default value is \emph{NA},\link{id2age}.}
}
\value{
Return 'F' means that the corresponding identity card for female citizens,'M' means male.
}
\description{
Extraction of gender in the resident identity cards of People's Republic of China.
}
\examples{

library(Ricetl)
x <- c('654003198111200241','341881197709275718',
'340503199401248097','469030199406204103','5111001')
id2gender(x)
id2gender(x,fillna=25)

}
\author{
Xu Jing
}
