% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Ricetl.R
\name{id2verify}
\alias{id2verify}
\title{To Determine Whether an ID Number is a Real ID Number}
\usage{
id2verify(x)
}
\arguments{
\item{x}{A vector containing 18 Bit ID numbers,\link{id2age}.}
}
\description{
To determine whether an ID number is a real ID number.
}
\examples{

library(Ricetl)
x <- c('654003198111200241','341881197709275718',
'340503199401248097','469030199406204103','51110019')
id2verify(x)

}
\author{
Xu Jing
}
