% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Ricetl.R
\docType{data}
\name{id_card6}
\alias{id_card6}
\title{Data Set for Address Correspondence}
\usage{
data(id_card6)
}
\description{
The form of the corresponding relationship between the
first six places of the ID card and the address of the household registration.
}
\examples{
library(Ricetl)
data(id_card6)
}
\author{
Xu Jing
}
