% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Ricetl.R
\name{Ricetl}
\alias{Ricetl}
\title{Data Extraction for People's Republic of China Resident Identity Card}
\usage{
Ricetl()
}
\description{
People's Republic of China resident identity card contains a lot of personal information.
It can extract the location of the household's residence in the identity card, the date of birth, the sex,
the location of the registered police station, and can identify the authenticity of the ID card through the check code.
Ricetl provides graphical user interface to process the function of extracting the above effective private information in the ID card number,
and provides a public data set related to the ID number.
}
\examples{

library(Ricetl)
library(gWidgetsRGtk2)
Ricetl()


}
\author{
Xu Jing
}
