% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Ricetl.R
\name{Ricetl}
\alias{Ricetl}
\title{R for Data Extraction of Resident Identity Card (PRC)}
\usage{
Ricetl()
}
\description{
Ricetl() method to extraction the resident identity card (PRC),Using Ricetl() dialytic ID data.
}
\examples{
\dontrun{
Ricetl()
}

}
\author{
XuJing
}
